Focus first client in the list when no client is in focus
If no client has the focus yet and the user performs a focus-next or focus-prev
action, the program was crashing because it attempted to focus a non-existent X
window.  There is now a check for this situation.  If no window is currently
in focus, the first window in the client list will be used.