Rework window auto focusing
Processing pointer motion events seems to be too expensive. (noticably slows
down the program)  The new approach is to poll the pointer location in the
event loop and periodically update the focus.  The poll is currently done
once every five iterations of the loop, which is definitely a hack but the
event loop as a whole needs to reworked so I'll probably tackle that next.
Also removed the old code for handling pointer motion events that was
already commented out.